+++ .eslintrc.js
@@ -20,7 +20,8 @@ module.exports = {
         custom: {
           regex: '^I[A-Z]',
           match: true
-        }
+        },
+        'prettier/prettier': ['error', { endOfLine: 'auto' }]
       }
     ],
     '@typescript-eslint/no-unused-vars': ['warn', { args: 'none' }],
