+++ .gitignore
@@ -37,6 +37,7 @@ share/python-wheels/
 .installed.cfg
 *.egg
 MANIFEST
+.yarn
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
