+++ package.json
@@ -68,7 +68,7 @@
     "react-hook-form": "^7.39.3",
     "react-leaflet": "~2.8.0",
     "react-leaflet-draw": "0.19.0",
-    "react-loader-spinner": "^5.3.4",
+    "react-loader-spinner": "^6.1.6",
     "react-modal": "3.15.1",
     "react-select": "5.4.0",
     "react-tooltip": "~5.26.3",
@@ -78,9 +78,9 @@
     "@jupyterlab/builder": "^3.5.2",
     "@types/classnames": "^2.3.1",
     "@types/isomorphic-fetch": "^0.0.36",
-    "@types/leaflet": "1.8.0",
+    "@types/leaflet": "^1.9.14",
     "@types/leaflet-draw": "^1.0.5",
-    "@types/lodash": "^4.14.185",
+    "@types/lodash": "^4.17.13",
     "@types/luxon": "^3.0.1",
     "@types/react": "^17.0.50",
     "@types/react-datepicker": "4.4.2",
@@ -98,11 +98,7 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
     "rimraf": "^3.0.2",
-    "typescript": "4.3.5"
-  },
-  "resolutions": {
-    "**/@types/react": "~17.0.50",
-    "**/@types/react-dom": "~17.0.17"
+    "typescript": "^5.7.2"
   },
   "jupyterlab": {
     "extension": true,
